Student Save opertaion done slight missing points
The missin point will be coverded by JQuery

+++ UniversityManagementSystem/UniversityManagementSystem/Models/EntityModel/Course.cs
@@ -12,7 +12,7 @@ namespace UniversityManagementSystem.Models
         public int Id { get; set; }
         public string Code { get; set; }
         public string Name { get; set; }
-        public decimal Credit { get; set; }
+        public double Credit { get; set; }
         public string Description { get; set; }
 
         public int DepartmentId { get; set; }
